Refactoring SampleData class, fixing style issues.
